requirements.txt
asgiref==3.7.2
Django==4.2.8
django-rest-framework==0.1.0
djangorestframework==3.14.0
install==1.3.5
psycopg==3.1.16
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.9.0